+++ src/Fable.Transforms/Python/Replacements.fs
@@ -1503,6 +1503,10 @@ let strings (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr opt
         | [ ExprType Char; ExprType(Number(Int32, NumberInfo.Empty)) ]
         | [ ExprType String; ExprType(Number(Int32, NumberInfo.Empty)) ] ->
             Helper.InstanceCall(c, "find", t, args, i.SignatureArgTypes, ?loc = r) |> Some
+        | [ ExprType String; StringComparisonEnumValue ] ->
+            Helper.InstanceCall(c, "find", t, [ List.head args ], i.SignatureArgTypes, ?loc = r) |> Some
+        | [ ExprType String; ExprType(Number(Int32, NumberInfo.Empty)); StringComparisonEnumValue ] ->
+            Helper.InstanceCall(c, "find", t, List.take 2 args, i.SignatureArgTypes, ?loc = r) |> Some
         | _ ->
             "The only extra argument accepted for String.IndexOf/LastIndexOf is startIndex."
             |> addErrorAndReturnNull com ctx.InlinePath r
@@ -1513,6 +1517,22 @@ let strings (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr opt
         | [ ExprType String ] -> Helper.InstanceCall(c, "rfind", t, args, i.SignatureArgTypes, ?loc = r) |> Some
         | [ ExprType Char as str; ExprType(Number(Int32, NumberInfo.Empty)) as start ]
         | [ ExprType String as str; ExprType(Number(Int32, NumberInfo.Empty)) as start ] ->
+            Helper.InstanceCall(
+                c,
+                "rfind",
+                t,
+                [
+                    str
+                    Value(NumberConstant(NumberValue.Int32 0, NumberInfo.Empty), None)
+                    start
+                ],
+                i.SignatureArgTypes,
+                ?loc = r
+            )
+            |> Some
+        | [ ExprType String; StringComparisonEnumValue ] ->
+            Helper.InstanceCall(c, "rfind", t, [ List.head args ], i.SignatureArgTypes, ?loc = r) |> Some
+        | [ ExprType String as str; ExprType(Number(Int32, NumberInfo.Empty)) as start; StringComparisonEnumValue ] ->
             Helper.InstanceCall(
                 c,
                 "rfind",